refactor: consolidate debt simulations
